+++ drivers/ethernet/dsa/Kconfig
@@ -67,4 +67,11 @@ config DSA_SPI
 	help
 	  Use SPI bus to communicate with PHY
 
+config DSA_NXP_IMX_NETC
+	bool "Support for NXP i.MX NETC"
+	default y
+	depends on DT_HAS_NXP_NETC_SWITCH_ENABLED
+	help
+	  Add support for NXP i.MX NETC DSA device driver.
+
 endif # DSA_DRIVERS
