+++ drivers/ethernet/nxp_imx_netc/CMakeLists.txt
@@ -6,5 +6,3 @@ if(CONFIG_ETH_NXP_IMX_NETC)
   zephyr_library_sources(eth_nxp_imx_netc_psi.c)
   zephyr_library_sources_ifdef(CONFIG_DT_HAS_NXP_IMX_NETC_BLK_CTRL_ENABLED eth_nxp_imx_netc_blk.c)
 endif()
-
-zephyr_library_sources_ifdef(CONFIG_DSA_NXP_IMX_NETC	dsa_nxp_imx_netc.c)
