+++ drivers/ethernet/nxp_imx_netc/Kconfig
@@ -99,14 +99,3 @@ config ETH_NXP_IMX_RX_RING_BUF_SIZE
 	  store one complete Ethernet frame, and be a multiple of 8.
 
 endif # ETH_NXP_IMX_NETC
-
-if NET_DSA
-
-config DSA_NXP_IMX_NETC
-	bool "Support for NXP i.MX NETC"
-	default y
-	depends on DT_HAS_NXP_NETC_SWITCH_ENABLED
-	help
-	  Add support for NXP i.MX NETC DSA device driver.
-
-endif
