+++ include/zephyr/drivers/ethernet/nxp_imx_netc.h
